containers/training/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.7.0
PyYAML>=6.0
ultralytics>=8.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
mlflow>=2.0.0
boto3>=1.28.0
sagemaker>=2.190.0
